Moved template match functionality to its own controller, and made logEvent a global util
