+++ lib/util/index.js
@@ -160,6 +160,17 @@ if (!Object.entries)
 };
 
 
+global.logEvent = function(path, label, msg) {
+  debug(path, label, msg);
+
+  let event = {};
+  event.path = path;
+  event.label = label;
+  event.msg = msg;
+
+  logger.info(event);
+}
+
 global.escapeRegExp = function(string) {
   return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
 }
